cloud_functions/raw-extract-github-commits/requirements.txt
functions-framework==3.*
requests
google-cloud-storage
google-cloud-secret-manager
